- Seperate app data management into a custom hook
