+++ src/components/Application.js
@@ -1,10 +1,10 @@
-import React, { useState, useEffect } from 'react';
-import axios from 'axios';
+import React from 'react';
 
 import 'components/Application.scss';
 import DayList from 'components/DayList';
 import Appointment from './Appointment';
 import { getAppointmentsForDay, getInterviewersForDay, getInterview } from 'helpers/selectors';
+import useApplicationData from 'hooks/useApplicationData';
 
 /**
  *
@@ -13,83 +13,19 @@ import { getAppointmentsForDay, getInterviewersForDay, getInterview } from 'help
 "GET_APPOINTMENTS": http://localhost:8001/api/appointments,
 "GET_INTERVIEWERS": http://localhost:8001/api/interviewers,
  */
-
 export default function Application() {
-  const [state, setState] = useState({
-    selectedDay: 'Monday',
-    days: [],
-    appointments: {},
-    interviewers: {},
-  });
-
+  const {
+    state,
+    setSelectedDay,
+    bookInterview,
+    cancelInterview
+  } = useApplicationData();
 
   const dailyAppointments = getAppointmentsForDay(state, state.selectedDay);
   const dailyInterviewers = getInterviewersForDay(state, state.selectedDay);
 
-  const setSelectedDay = (selectedDay) => {
-    setState({ ...state, selectedDay });
-  };
-
-  const bookInterview = (id, interview) => {
-    const appointment = {
-      ...state.appointments[id],
-      interview: { ...interview }
-    };
-
-    const appointments = {
-      ...state.appointments,
-      [id]: appointment
-    };
-
-    return axios.put(`/api/appointments/${id}`, { interview })
-      .then((res) => {
-        setState({ ...state, appointments });
-        return res;
-      }).catch(error => {
-        throw error;
-      });
-  };
-
-  const cancelInterview = (appointmentId) => {
-    return axios.delete(`/api/appointments/${appointmentId}`)
-      .then((res) => {
-        console.log('res = ', res);
-        setState(prev => {
-          // const { interview } = prev.appointments.find(id => id === appointmentId);
-          return {
-            ...prev,
-            appointments: { ...prev.appointments, [appointmentId]: { ...prev.appointments[appointmentId], interview: null } },
-          };
-        });
-        return res;
-      }).catch(error => {
-        throw error;
-      });
-  };
-
-  useEffect(() => {
-    Promise.all([
-      axios.get('/api/days'),
-      axios.get('/api/appointments'),
-      axios.get('/api/interviewers'),
-    ])
-      .then((responses) => {
-        // console.log(responses);
-        setState((prev) => ({
-          ...prev,
-          days: responses[0].data,
-          appointments: responses[1].data,
-          interviewers: responses[2].data,
-        }));
-      })
-      .catch((error) => {
-        console.log(error);
-      });
-  }, []);
-
-  const schedule = dailyAppointments.map((appointment) => {
+  const appointmentsComponents = dailyAppointments.map((appointment) => {
     const interview = getInterview(state, appointment.interview);
-
     return (
       <Appointment
         key={appointment.id}
@@ -126,7 +62,7 @@ export default function Application() {
         />
       </section>
       <section className='schedule'>
-        {schedule}
+        {appointmentsComponents}
         <Appointment key='last' time='5pm' /> {/* Do not forget this */}
       </section>
     </main>
